Add JSON output via the API and CLI
This is the important part of a WASM port. If you're using this project as a WAS
M port, chances are whatever API you pipe the output to doesn't support ANSI color codes. In my case, the Canvas element certainly does not. This JSON output via the API gives us the option to output a JSON format array (structured as {char: string, rgb: uint[4]}) and use RGB